Update Set C,a.java
 Create an interface Department containing attributes deptName and deptHead. It also
has abstract methods for printing the attributes. Create a class hostel containing
hostelName, hostelLocation and numberOfRooms. The class contains method printing
the attributes. Then write Student class extending the Hostel class and implementing
the Department interface. This class contains attributes studentName, regNo,
electiveSubject and avgMarks. Write suitable printData method for this class. Also,
implement the abstract methods of the Department interface. Write a driver class to
test the Student class. The program should be menu driven containing the options:
i. Admit new student
ii. Migrate a student
iii. Display details of a student
For the third option, a search is to be made on the basis of the entered registration
Number.